requirements.txt
fastapi
uvicorn
python-telegram-bot
aiohttp
pyyaml
psutil
openai-whisper
piper-tts
python-dotenv